renamed some stuff, tests
