+++ .gitignore
@@ -22,6 +22,6 @@ ui.js
 ui.css
 ui.md
 .vscode
-py/__pycache__/*
-py/.idea/
-py/db
+flask_api/__pycache__/*
+flask_api/.idea/
+flask_api/db
