+++ requirements.txt
@@ -7,15 +7,22 @@ cligj==0.7.2
 fiona==1.9.5
 Flask==3.0.0
 Flask-Cors==4.0.0
+Flask-JWT-Extended==4.5.3
 geopandas==0.14.0
+gunicorn==21.2.0
+iniconfig==2.0.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
 MarkupSafe==2.1.3
 numpy==1.26.2
 packaging==23.2
 pandas==2.1.3
+pluggy==1.3.0
+PyJWT==2.8.0
 pyproj==3.6.1
+pytest==7.4.3
 python-dateutil==2.8.2
+python-dotenv==1.0.0
 pytz==2023.3.post1
 shapely==2.0.2
 six==1.16.0
